Change the folder structue. Added new routing class to handle Benfords related routes.
