+++ src/main/kotlin/util/Constants.kt
@@ -1,6 +1,7 @@
 package com.example.util
 
 object Constants {
-    const val BASE_PATH_ = "/api/v1/"
+    const val BASE_PATH_ = "/api"
+    const val V1_PATH_ = "/v1"
     const val BENFORDS_ANALYSIS_PATH="/analysis/benfords"
 }
